final notes

+++ routes/web.php
@@ -1,9 +1,11 @@
 <?php
 
-use App\Http\Controllers\ListingController;
+use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Models\Listing;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
+
 
 //@1ROUTING & RESPONSES*
 
@@ -11,25 +13,45 @@ use App\Models\Listing;
 Route::get('/', [ListingController::class, 'index']);
 
 //Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
-//Store Listing
-Route::post('/listings', [ListingController::class, 'store']);
+
+
+//Store Listing data
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 
 //Show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
+//manage listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 //Single Listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
+
+//Show Register/Create Form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+//Create new user
+Route::post('/users', [UserController::class, 'store']);
+
+
+//Logout
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+//Show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+//login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
 
 
 // Common Resource Routes:
